# contracts/http/openapi.yaml
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "EcoBin P0 HTTP API",
    "version": "1.0.0-rc.1",
    "description": "EcoBin /api/v1 transport contract. Business slices add their own paths and schemas without redefining these cross-cutting rules."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin deployment"
    }
  ],
...
  "paths": {
...
    "/api/v1/web/platform/tenants/{tenantCode}": {
      "parameters": [{"$ref": "#/components/parameters/TenantCode"}],
      "get": {
        "operationId": "getIdentityTenant",
        "tags": ["Identity directory"],
        "summary": "Read one tenant through the platform privilege boundary",
        "security": [{"webSessionCookie": []}],
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"}
        }
      }
    },
...
    "/api/v1/web/platform/tenants/{tenantCode}/organizations/{organizationCode}": {
      "parameters": [
        {"$ref": "#/components/parameters/TenantCode"},
        {"$ref": "#/components/parameters/OrganizationCode"}
      ],
      "get": {
        "operationId": "getPlatformIdentityOrganization",
        "tags": ["Identity directory"],
        "summary": "Read an organization in one explicit platform target tenant",
        "security": [{"webSessionCookie": []}],
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"}
        }
      }
    },
...
    "/api/v1/web/organizations/{organizationCode}": {
      "parameters": [{"$ref": "#/components/parameters/OrganizationCode"}],
      "get": {
        "operationId": "getCurrentTenantOrganization",
        "tags": ["Identity directory"],
        "summary": "Read one visible organization in the session tenant",
        "security": [{"webSessionCookie": []}],
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"}
        }
      }
    },
...
    "/api/v1/web/organizations/{organizationCode}/organization-users/{organizationUserUid}": {
      "get": {
        "operationId": "getCurrentTenantOrganizationUser",
        "tags": ["Identity directory"],
        "summary": "Read one safe organization-user projection",
        "security": [{"webSessionCookie": []}],
        "parameters": [
          {"$ref": "#/components/parameters/OrganizationCode"},
          {"$ref": "#/components/parameters/OrganizationUserUid"}
        ],
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "403": {"$ref": "#/components/responses/ForbiddenProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"}
        }
      }
    },
    "/api/v1/web/organizations/{organizationCode}/organization-users/{organizationUserUid}/freezes": {
      "post": {
        "operationId": "freezeCurrentTenantOrganizationUser",
        "tags": ["Identity directory"],
        "summary": "Freeze one organization user and revoke ordinary miniapp sessions",
        "security": [{"webSessionCookie": [], "webCsrfToken": []}],
        "parameters": [
          {"$ref": "#/components/parameters/OrganizationCode"},
          {"$ref": "#/components/parameters/OrganizationUserUid"},
          {"$ref": "#/components/parameters/IdempotencyKey"}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccountVersionCommand"}}}
        },
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "400": {"$ref": "#/components/responses/InvalidRequest"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "403": {"$ref": "#/components/responses/ForbiddenProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"},
          "409": {"$ref": "#/components/responses/ConflictProblem"},
          "422": {"$ref": "#/components/responses/BusinessRuleProblem"}
        }
      }
    },
    "/api/v1/web/organizations/{organizationCode}/organization-users/{organizationUserUid}/restorations": {
      "post": {
        "operationId": "restoreCurrentTenantOrganizationUser",
        "tags": ["Identity directory"],
        "summary": "Restore one frozen organization user without restoring old sessions",
        "security": [{"webSessionCookie": [], "webCsrfToken": []}],
        "parameters": [
          {"$ref": "#/components/parameters/OrganizationCode"},
          {"$ref": "#/components/parameters/OrganizationUserUid"},
          {"$ref": "#/components/parameters/IdempotencyKey"}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccountVersionCommand"}}}
        },
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "400": {"$ref": "#/components/responses/InvalidRequest"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "403": {"$ref": "#/components/responses/ForbiddenProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"},
          "409": {"$ref": "#/components/responses/ConflictProblem"},
          "422": {"$ref": "#/components/responses/BusinessRuleProblem"}
        }
      }
    },
    "/api/v1/web/organizations/{organizationCode}/organization-users/{organizationUserUid}/capabilities/clean-operation/grants": {
      "post": {
        "operationId": "grantCurrentTenantOrganizationUserCleanOperation",
        "tags": ["Identity directory"],
        "summary": "Grant the organization user's cleaning capability and revoke ordinary miniapp sessions",
        "security": [{"webSessionCookie": [], "webCsrfToken": []}],
        "parameters": [
          {"$ref": "#/components/parameters/OrganizationCode"},
          {"$ref": "#/components/parameters/OrganizationUserUid"},
          {"$ref": "#/components/parameters/IdempotencyKey"}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccountVersionCommand"}}}
        },
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "400": {"$ref": "#/components/responses/InvalidRequest"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "403": {"$ref": "#/components/responses/ForbiddenProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"},
          "409": {"$ref": "#/components/responses/ConflictProblem"},
          "422": {"$ref": "#/components/responses/BusinessRuleProblem"}
        }
      }
    },
    "/api/v1/web/organizations/{organizationCode}/organization-users/{organizationUserUid}/capabilities/clean-operation/revocations": {
      "post": {
        "operationId": "revokeCurrentTenantOrganizationUserCleanOperation",
        "tags": ["Identity directory"],
        "summary": "Revoke the organization user's cleaning capability and revoke ordinary miniapp sessions",
        "security": [{"webSessionCookie": [], "webCsrfToken": []}],
        "parameters": [
          {"$ref": "#/components/parameters/OrganizationCode"},
          {"$ref": "#/components/parameters/OrganizationUserUid"},
          {"$ref": "#/components/parameters/IdempotencyKey"}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccountVersionCommand"}}}
        },
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "400": {"$ref": "#/components/responses/InvalidRequest"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "403": {"$ref": "#/components/responses/ForbiddenProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"},
          "409": {"$ref": "#/components/responses/ConflictProblem"},
          "422": {"$ref": "#/components/responses/BusinessRuleProblem"}
        }
      }
    },
...
    "/api/v1/web/platform/tenants/{tenantCode}/organizations/{organizationCode}/organization-users": {
      "get": {
        "operationId": "listPlatformOrganizationUsers",
        "tags": ["Identity directory"],
        "summary": "Platform list of safe organization-user projections",
        "security": [{"webSessionCookie": []}],
        "parameters": [
          {"$ref": "#/components/parameters/TenantCode"},
          {"$ref": "#/components/parameters/OrganizationCode"},
          {"$ref": "#/components/parameters/Page"},
          {"$ref": "#/components/parameters/PageSize"},
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {"type": "string", "enum": ["ACTIVE", "FROZEN"]}
          },
          {
            "name": "phoneBound",
            "in": "query",
            "required": false,
            "schema": {"type": "boolean"}
          },
          {
            "name": "registeredFrom",
            "in": "query",
            "required": false,
            "schema": {"$ref": "#/components/schemas/UtcTimestamp"}
          },
          {
            "name": "registeredTo",
            "in": "query",
            "required": false,
            "schema": {"$ref": "#/components/schemas/UtcTimestamp"}
          },
          {
            "name": "sourceDeploymentCode",
            "in": "query",
            "required": false,
            "schema": {"type": "string", "minLength": 1, "maxLength": 64}
          },
          {
            "name": "cleanOperation",
            "in": "query",
            "required": false,
            "schema": {"type": "boolean"}
          }
        ],
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "400": {"$ref": "#/components/responses/InvalidRequest"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"}
        }
      }
    },
    "/api/v1/web/platform/tenants/{tenantCode}/organizations/{organizationCode}/organization-users/{organizationUserUid}": {
      "get": {
        "operationId": "getPlatformOrganizationUser",
        "tags": ["Identity directory"],
        "summary": "Platform read of one safe organization-user projection",
        "security": [{"webSessionCookie": []}],
        "parameters": [
          {"$ref": "#/components/parameters/TenantCode"},
          {"$ref": "#/components/parameters/OrganizationCode"},
          {"$ref": "#/components/parameters/OrganizationUserUid"}
        ],
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"}
        }
      }
    },
    "/api/v1/web/platform/tenants/{tenantCode}/organizations/{organizationCode}/organization-users/{organizationUserUid}/freezes": {
      "post": {
        "operationId": "freezePlatformOrganizationUser",
        "tags": ["Identity directory"],
        "summary": "Platform freeze of one organization user",
        "security": [{"webSessionCookie": [], "webCsrfToken": []}],
        "parameters": [
          {"$ref": "#/components/parameters/TenantCode"},
          {"$ref": "#/components/parameters/OrganizationCode"},
          {"$ref": "#/components/parameters/OrganizationUserUid"},
          {"$ref": "#/components/parameters/IdempotencyKey"}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccountVersionCommand"}}}
        },
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "400": {"$ref": "#/components/responses/InvalidRequest"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"},
          "409": {"$ref": "#/components/responses/ConflictProblem"},
          "422": {"$ref": "#/components/responses/BusinessRuleProblem"}
        }
      }
    },
    "/api/v1/web/platform/tenants/{tenantCode}/organizations/{organizationCode}/organization-users/{organizationUserUid}/restorations": {
      "post": {
        "operationId": "restorePlatformOrganizationUser",
        "tags": ["Identity directory"],
        "summary": "Platform restore of one frozen organization user",
        "security": [{"webSessionCookie": [], "webCsrfToken": []}],
        "parameters": [
          {"$ref": "#/components/parameters/TenantCode"},
          {"$ref": "#/components/parameters/OrganizationCode"},
          {"$ref": "#/components/parameters/OrganizationUserUid"},
          {"$ref": "#/components/parameters/IdempotencyKey"}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccountVersionCommand"}}}
        },
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "400": {"$ref": "#/components/responses/InvalidRequest"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"},
          "409": {"$ref": "#/components/responses/ConflictProblem"},
          "422": {"$ref": "#/components/responses/BusinessRuleProblem"}
        }
      }
    },
    "/api/v1/web/platform/tenants/{tenantCode}/organizations/{organizationCode}/organization-users/{organizationUserUid}/capabilities/clean-operation/grants": {
      "post": {
        "operationId": "grantPlatformOrganizationUserCleanOperation",
        "tags": ["Identity directory"],
        "summary": "Platform grant of one organization user's cleaning capability",
        "security": [{"webSessionCookie": [], "webCsrfToken": []}],
        "parameters": [
          {"$ref": "#/components/parameters/TenantCode"},
          {"$ref": "#/components/parameters/OrganizationCode"},
          {"$ref": "#/components/parameters/OrganizationUserUid"},
          {"$ref": "#/components/parameters/IdempotencyKey"}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccountVersionCommand"}}}
        },
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "400": {"$ref": "#/components/responses/InvalidRequest"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"},
          "409": {"$ref": "#/components/responses/ConflictProblem"},
          "422": {"$ref": "#/components/responses/BusinessRuleProblem"}
        }
      }
    },
    "/api/v1/web/platform/tenants/{tenantCode}/organizations/{organizationCode}/organization-users/{organizationUserUid}/capabilities/clean-operation/revocations": {
      "post": {
        "operationId": "revokePlatformOrganizationUserCleanOperation",
        "tags": ["Identity directory"],
        "summary": "Platform revocation of one organization user's cleaning capability",
        "security": [{"webSessionCookie": [], "webCsrfToken": []}],
        "parameters": [
          {"$ref": "#/components/parameters/TenantCode"},
          {"$ref": "#/components/parameters/OrganizationCode"},
          {"$ref": "#/components/parameters/OrganizationUserUid"},
          {"$ref": "#/components/parameters/IdempotencyKey"}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccountVersionCommand"}}}
        },
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "400": {"$ref": "#/components/responses/InvalidRequest"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"},
          "409": {"$ref": "#/components/responses/ConflictProblem"},
          "422": {"$ref": "#/components/responses/BusinessRuleProblem"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "webSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-ecobin-web-session",
        "description": "Secure, HttpOnly, SameSite=Lax, Path=/ and no Domain. JavaScript never reads this value."
      },
      "webCsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-TOKEN",
        "description": "Bootstrapped from GET /api/v1/web/auth/csrf-token and rotated after login/logout."
      },
      "miniappBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Exactly one short-lived aud=miniapp or aud=miniapp-staff session. No refresh token."
      },
      "wechatPaySignature": {
        "type": "apiKey",
        "in": "header",
        "name": "Wechatpay-Signature"
      },
      "wechatPayTimestamp": {
        "type": "apiKey",
        "in": "header",
        "name": "Wechatpay-Timestamp"
      },
      "wechatPayNonce": {
        "type": "apiKey",
        "in": "header",
        "name": "Wechatpay-Nonce"
      },
      "wechatPaySerial": {
        "type": "apiKey",
        "in": "header",
        "name": "Wechatpay-Serial"
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "UUIDv4 generated once for one human intent and reused by every retry of that same intent.",
        "schema": {
          "$ref": "#/components/schemas/UuidV4"
        }
      },
      "TenantCode": {
        "name": "tenantCode",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/TenantCode"
        }
      },
      "OrganizationCode": {
        "name": "organizationCode",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/OrganizationCode"
        }
      },
...
      "OrganizationUserUid": {
        "name": "organizationUserUid",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/PublicUid"
        }
      },
...
      "Page": {
        "name": "page",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 200,
          "default": 20
        }
      },
...
    },
    "headers": {
      "NoStore": {
        "description": "Sensitive or mutable projections are not cacheable",
        "schema": {
          "type": "string",
          "const": "no-store"
        }
      },
      "RequestId": {
        "description": "Transport request identifier; never a business idempotency identity",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128
        }
      },
...
    },
    "responses": {
      "DirectoryOk": {
        "description": "V-01 identity directory projection returned from current database facts",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          },
          "X-Request-Id": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DirectoryEnvelope"
            }
          }
        }
      },
...
      "InvalidRequest": {
        "description": "Malformed JSON, type, format, or cross-field input",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      },
      "UnauthorizedProblem": {
        "description": "Authentication is missing, invalid, revoked, expired, or has the wrong audience",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      },
      "ForbiddenProblem": {
        "description": "The visible resource exists but the authenticated subject lacks the current capability",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      },
      "NotFoundProblem": {
        "description": "The resource does not exist or is outside the caller's current visible scope",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      },
...
      "ConflictProblem": {
        "description": "Idempotency digest, expected version, state, or occupancy conflict",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      },
      "BusinessRuleProblem": {
        "description": "The input is valid but a current business qualification or invariant rejects it",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      },
...
    },
    "schemas": {
      "TenantCode": {
        "type": "string",
        "minLength": 1,
        "maxLength": 32,
        "pattern": "^[a-z0-9][a-z0-9-]*$"
      },
      "OrganizationCode": {
        "type": "string",
        "minLength": 1,
        "maxLength": 32,
        "pattern": "^[a-z0-9][a-z0-9-]*$"
      },
...
      "DirectoryStatus": {
        "type": "string",
        "enum": [
          "ENABLED",
          "DISABLED"
        ]
      },
...
      "IdentityTenant": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "tenantCode",
          "enterpriseName",
          "status",
          "version",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "tenantCode": {
            "$ref": "#/components/schemas/TenantCode"
          },
          "enterpriseName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "status": {
            "$ref": "#/components/schemas/DirectoryStatus"
          },
          "contactName": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 100
          },
          "contactPhone": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 32
          },
          "contactAddress": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 500
          },
          "version": {
            "$ref": "#/components/schemas/ExpectedVersion"
          },
          "principalAccount": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/PrincipalAccountSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "$ref": "#/components/schemas/UtcTimestamp"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/UtcTimestamp"
          }
        }
      },
...
      "ProvisionedOrganizationStaff": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "staffAccount",
          "membership"
        ],
        "properties": {
          "staffAccount": {
            "$ref": "#/components/schemas/StaffAccount"
          },
          "membership": {
            "$ref": "#/components/schemas/OrganizationMembership"
          }
        }
      },
...
      "DirectoryEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "data",
          "requestId"
        ],
        "properties": {
          "code": {
            "type": "string",
            "const": "OK"
          },
          "data": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/IdentityTenant"
              },
              {
                "$ref": "#/components/schemas/IdentityOrganization"
              },
              {
                "$ref": "#/components/schemas/StaffAccount"
              },
              {
                "$ref": "#/components/schemas/OrganizationMembership"
              },
              {
                "$ref": "#/components/schemas/EffectiveAccess"
              },
              {
                "$ref": "#/components/schemas/ProvisionedOrganizationStaff"
              },
              {
                "$ref": "#/components/schemas/OrganizationUserLookup"
              },
              {
                "$ref": "#/components/schemas/OrganizationUser"
              },
              {
                "$ref": "#/components/schemas/StaffMiniappBindingLookup"
              },
              {
                "$ref": "#/components/schemas/StaffMiniappBinding"
              },
              {
                "$ref": "#/components/schemas/DirectoryPage"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/PermissionDefinition"
                }
              }
            ]
          },
          "requestId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        }
      },
...
      "AccountVersionCommand": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "expectedVersion",
          "expectedAuthVersion"
        ],
        "properties": {
          "expectedVersion": {
            "$ref": "#/components/schemas/ExpectedVersion"
          },
          "expectedAuthVersion": {
            "$ref": "#/components/schemas/ExpectedVersion"
          },
          "reason": {
            "type": "string",
            "maxLength": 500
          }
        }
      },
...
      "UuidV4": {
        "type": "string",
        "format": "uuid",
        "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
        "example": "4b9e8c22-7d38-4e4b-a1b0-7a0e1656e001"
      },
      "PublicUid": {
        "allOf": [
          {
            "$ref": "#/components/schemas/UuidV4"
          }
        ],
        "description": "Stable public identity; never an internal BIGINT primary key"
      },
...
      "UtcTimestamp": {
        "type": "string",
        "format": "date-time",
        "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}\\.[0-9]{3}Z$",
        "example": "2026-07-26T03:15:20.123Z"
      },
...
      "ExpectedVersion": {
        "type": "integer",
        "minimum": 0,
        "description": "Last server version explicitly observed by the operator",
        "example": 3
      },
...
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "requestId",
          "retryable",
          "details"
        ],
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]*(\\.[A-Z][A-Z0-9_]*)+$",
            "example": "COMMON.IDEMPOTENCY_KEY_CONFLICT"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500,
            "example": "相同操作标识已绑定到不同请求"
          },
          "requestId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "example": "01J3EXAMPLE00000000000001"
          },
          "retryable": {
            "type": "boolean",
            "example": false
          },
          "details": {
            "type": "object",
            "additionalProperties": true,
            "example": {}
          }
        }
      },
...
    }
  },
...
}